Add tooltips to garden
This also collapses everything down to a single item type, using components to do the rest. This might be a bad idea, but it was easier to keep the one block item than trying to do the multiple items mapping to one block that the previous version of the mod had.